refactor: Add stats service to Fan.Blog #266
