+++ src/Fan.Blog/Data/IPostRepository.cs
@@ -1,5 +1,6 @@
 ﻿using Fan.Blog.Enums;
 using Fan.Blog.Models;
+using Fan.Blog.Stats;
 using Fan.Data;
 using System;
 using System.Collections.Generic;
