+++ src/Fan.Blog/Stats/IStatsService.cs
@@ -1,0 +1,21 @@
+﻿using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace Fan.Blog.Stats
+{
+    public interface IStatsService
+    {
+        /// <summary>
+        /// Returns a dictionary of year and months, the key is year and the value is a list of 
+        /// <see cref="MonthItem"/> objects.
+        /// </summary>
+        /// <returns></returns>
+        Task<Dictionary<int, List<MonthItem>>> GetArchivesAsync();
+
+        /// <summary>
+        /// Returns total number of posts by each <see cref="EPostStatus"/>.
+        /// </summary>
+        /// <returns></returns>
+        Task<PostCount> GetPostCountAsync();
+    }
+}
