+++ src/Fan.Web/Themes/Clarity/Views/_ViewImports.cshtml
@@ -2,6 +2,7 @@
 @using Fan.Blog.Models
 @using Fan.Blog.Categories
 @using Fan.Blog.Tags
+@using Fan.Blog.Stats
 @using Fan.Blog.ViewModels
 @using Fan.Blog.Services
 @using Fan.Membership
